# Make `x test tidy` less noisy